Animate focusing on a post when it is clicked
Using CSS animations instead of JS animations for speed.
Zepto is used to toggle the class on and off.

This isn't perfect, the animation only works when you toggle it on,
when you toggle it off it just jumps back to how it was.

Not sure if I should actually use this in an app...
Maybe a normal page redirect from 'index' to 'show' and then the
animation can happen on load for the 'show' page?